+++ mern_vb_backend/server.js
@@ -17,6 +17,7 @@ app.use('/api/payments', require('./routes/payment'));
 app.use('/api/transactions', require('./routes/transactions'));
 app.use('/api/savings', require('./routes/savings'));
 app.use('/api/loans', require('./routes/loans'));
+app.use('/api/cycle', require('./routes/cycle'));
 
 const clientOptions = {
   serverApi: {